pyproject.toml
[project]
name = "yazr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bigtree>=0.29.2",
    "cachetools>=6.1.0",
    "fabric>=3.2.2",
    "python-slugify>=8.0.4",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.3",
    "nbstripout>=0.8.1",
    "commitizen>=4.8.3",
    "ipdb>=0.13.13",
    "ipython>=9.2.0",
    "nbdime>=4.0.2",
    "pre-commit>=4.2.0",
    "ruff>=0.11.11",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "types-python-slugify>=8.0.2.20240310",
]

[tool.uv]
default-groups = []

[tool.uv.workspace]
members = ["yazr"]

[tool.ruff]
line-length = 88
indent-width = 4

target-version = "py311"

[tool.ruff.lint]

# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
# Ignore for now:
#   - F403 (star imports)
#   - E722 (bare excepts)
ignore = ["F403", "F405", "E722"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]

# use single quotes for strings.
quote-style = "double"

# indent with spaces, rather than tabs.
indent-style = "space"

# respect magic trailing commas.
skip-magic-trailing-comma = false

# automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.pydocstyle]
convention = "google"  # Accepts: "google", "numpy", or "pep257".


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
